# docs/registers/openapi.yaml
...
openapi: 3.0.3
info:
  title: RFS Station Manager API
  description: |
    REST API for the RFS Station Manager - a real-time digital sign-in system for Rural Fire Service stations.
    
    This API provides endpoints for:
    - Member management
    - Activity tracking
    - Check-in/check-out operations
    - Event management
    - Truck checks
    - Achievements
    
    Real-time updates are also available via Socket.io WebSocket connections.
  version: 1.0.0
  contact:
    name: RFS Station Manager Support
    url: https://github.com/richardthorek/Station-Manager
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://bungrfsstation.azurewebsites.net
    description: Production server
...
paths:
...
  /api/auth/signup:
    post:
      tags:
        - Organizations
      summary: Self-service sign-up with facility claim
      description: >
        Creates a Community-plan Organization and its owner account. Requires `email` and a
        `facility` selection — either `{facilityKey}` (first-come-first-served claim of a Digital
        Atlas facility) or `{custom: {name, serviceType, ...}}` (an unlisted unit, no review needed).
        Full request/response schema: docs/registers/api_register.json → auth["POST /api/auth/signup"].
      operationId: signup
      responses:
        '201':
          description: Account + organization created
        '400':
          description: Invalid input, unknown facilityKey, or missing custom facility details
        '409':
          description: Username taken, or code=FACILITY_ALREADY_CLAIMED (a ClaimConflict is recorded)
  /api/auth/switch-org:
    post:
      tags:
        - Organizations
      summary: Switch active organization (multi-org membership)
      description: >
        Reissues the JWT scoped to a different organization the caller is an active member of.
        Full schema: docs/registers/api_register.json → auth["POST /api/auth/switch-org"].
      operationId: switchOrg
      security:
        - bearerAuth: []
      responses:
        '200':
          description: New token + organization
        '403':
          description: Not a member of the target organization
...
components:
...
  securitySchemes:
    # Future: JWT Bearer token authentication for admin functions
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

# No security required currently (by design for kiosk access)
# Future versions may require authentication for admin operations
security: []